refactored and rearranged the ChargeDensitySource capability
