+++ Source/Input/MacroscopicProperties/MacroscopicProperties.cpp
@@ -1,13 +1,7 @@
-#include "MacroscopicProperties.H"
+#include "MacroscopicProperties_impl.H"
 
 #include "../../Utils/SelectWarpXUtils/WarpXUtil.H"
 #include "../../Utils/CodeUtils/CodeUtil.H"
-
-#include "Code.H"
-#include "GeometryProperties.H"
-
-#include <AMReX_ParmParse.H>
-#include <AMReX_Parser.H>
 
 #include <ctype.h>
 
@@ -170,7 +164,14 @@ c_MacroscopicProperties::InitData()
         DefineAndInitializeMacroparam(macro_str, macro_num, ba, dm, geom, Ncomp1, map_num_ghostcell[macro_str]);
     }
     
-    Deposit_ExternalChargeDensitySources();
+    Deposit_AllExternalChargeDensitySources();
+}
+
+
+void 
+c_MacroscopicProperties::Deposit_AllExternalChargeDensitySources()
+{
+    if (p_PointChargeSource) Deposit_ExternalChargeDensitySources(p_PointChargeSource);
 }
 
 
@@ -327,59 +328,9 @@ c_MacroscopicProperties::Define_ExternalChargeDensitySources()
             v_pointCharges.emplace_back(pos.data(), sigma, charge_unit);
         }
 
-        if (p_ChargeDensitySource == nullptr) {
+        if (!p_PointChargeSource) {
             bool print=true;
-            p_ChargeDensitySource = std::make_unique<PointChargeSource>(std::move(v_pointCharges), print);
+            p_PointChargeSource = std::make_unique<PointChargeSource>(std::move(v_pointCharges), print);
         }
-    }
-}
-
-
-void
-c_MacroscopicProperties::Deposit_ExternalChargeDensitySources()
-{
-    auto& rCode  = c_Code::GetInstance();
-    auto& rGprop = rCode.get_GeometryProperties();
-    auto& geom = rGprop.geom;
-    auto dx = geom.CellSizeArray();
-    auto& real_box = geom.ProbDomain();
-
-    amrex::MultiFab* const p_charge_density_mf = get_p_mf("charge_density");
-    
-    if(p_charge_density_mf == nullptr) 
-        AMREX_ALWAYS_ASSERT_WITH_MESSAGE(p_charge_density_mf != nullptr, 
-                "charge_density mf string wrong!");
-
-
-    auto iv = p_charge_density_mf->ixType().toIntVect();
-
-    for ( amrex::MFIter mfi(*p_charge_density_mf, amrex::TilingIfNotGPU()); 
-                        mfi.isValid(); 
-                        ++mfi )
-    {
-        const auto& tb = mfi.tilebox( iv, p_charge_density_mf->nGrowVect() );
-        auto const& mf_array = p_charge_density_mf->array(mfi);
-
-        int num_sources = p_ChargeDensitySource->get_num_sources();
-        const auto* p_vec_source = p_ChargeDensitySource->get_p_sources();
-        auto charge_density_calculator = p_ChargeDensitySource->get_charge_density_calculator();
-
-        amrex::ParallelFor (tb,
-            [=] 
-            AMREX_GPU_DEVICE (int i, int j, int k)
-        {
-                amrex::Real fac_x = (1._rt - iv[0]) * dx[0] * 0.5_rt;
-                amrex::Real x = i * dx[0] + real_box.lo(0) + fac_x;
-
-                amrex::Real fac_y = (1._rt - iv[1]) * dx[1] * 0.5_rt;
-                amrex::Real y = j * dx[1] + real_box.lo(1) + fac_y;
-
-                amrex::Real fac_z = (1._rt - iv[2]) * dx[2] * 0.5_rt;
-                amrex::Real z = k * dx[2] + real_box.lo(2) + fac_z;
-
-                for(int s=0; s< num_sources; ++s) {
-                    mf_array(i,j,k) += charge_density_calculator(p_vec_source[s],x,y,z);
-                }
-        });
     }
 }
